refactor(webApi): update the webapi utils file - Update the webAPI file to accept other function calls [Finishes #137598387]
